% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptron.R
\name{imptron}
\alias{imptron}
\title{Imperceptron matrix}
\usage{
imptron(x, act = NULL, param = c(rep(1, ncol(x)), rep(0, ncol(x))), ...)
}
\arguments{
\item{x}{Matrix of input values.}

\item{act}{Optional activation function to pass through after summing weights and bias. See functions included in \link{activations} or provide your own.}

\item{param}{Vector of weights and biases to be optimized (e.g., using \link{optim}). The first \code{ncol(x)} are weights, and the remaining
are biases. This is normally taken from the output of \link{imptron_train}.}

\item{...}{Additional arguments to pass to the activation function.}
}
\value{
Matrix of same dimensions as \code{x}.
}
\description{
Calculate values at a layer of perceptrons that match dimensions of the input matrix, with no interaction.
}
\details{
This function multiplies each row of an input matrix by a vector of weights, plus a vector of biases, and optionally passes the output through
an activation function (such as found in \link{activations}). These parameters can be taken from the output of \link{imptron_train}.
}
