% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptron_train.R
\name{imptron_train}
\alias{imptron_train}
\title{Train an imperceptron matrix mapper}
\usage{
imptron_train(
  x,
  y,
  method = "Nelder-Mead",
  act = NULL,
  optim_control = list(trace = 1, maxit = 5000),
  ...
)
}
\arguments{
\item{x}{Input matrix.}

\item{y}{Target matrix.}

\item{method}{Optimization method, most likely from \link{optim}.}

\item{act}{Optional activation function to pass through after summing weights and biases of perceptrons. See functions included in \link{activations} or provide your own.}

\item{...}{Additional arguments to pass to the activation function.}

\item{oprim_control}{List of control arguments to pass to \link{optim}.}
}
\value{
A list with the \link{optim} output and other arguments specified in the call.
}
\description{
Another wrapper for \link{optim}. This one optimizes the mapping between two matrices without any interaction between perceptrons until the
final activation function, if specified.
}
\details{
Optimize weights and biases of non-dense perceptrons to map input matrix \code{x} to \code{y}. There are no dense hidden units.
}
