% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptron.R
\name{ptron}
\alias{ptron}
\title{Perceptron matrix}
\usage{
ptron(
  x,
  units = ncol(x),
  act = NULL,
  param = c(rep(1, ncol(x) * units), rep(0, units)),
  ...
)
}
\arguments{
\item{x}{Matrix of input values.}

\item{units}{Vector. Number of hidden densely connected perceptron units.}

\item{act}{Optional activation function to pass through after summing weights and bias. See functions included in \link{activations} or provide your own.}

\item{param}{Vector of weights and biases to be optimized (e.g., using \link{optim}). The first \code{ncol(mat) * units} are weights, and the remaining
are biases for each \code{unit}. This is normally tkaen from the output of \link{ptron_train}.}

\item{...}{Additional arguments to pass to the activation function.}
}
\value{
A matrix of dimensions \code{nrow(x)} by \code{length(units)}.
}
\description{
Calculate the values at a layer of perceptrons.
}
\details{
This function multiplies each row of a matrix of input values by a matrix of weights, adds a vector of biases, and optionally passes the output through
an activation function (such as found in \link{activations}). These parameters can be taken from the output of \link{ptron_train}.
}
