% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptron_train.R
\name{ptron_train}
\alias{ptron_train}
\title{Train a perceptron matrix mapper}
\usage{
ptron_train(
  x,
  y,
  units = ncol(x),
  method = "Nelder-Mead",
  optim_control = list(trace = 1, maxit = 5000),
  act = NULL,
  ...
)
}
\arguments{
\item{x}{Input matrix.}

\item{y}{Target matrix.}

\item{units}{Vector. Number of hidden densely connected perceptron units.}

\item{method}{Optimization method, most likely from \link{optim}.}

\item{act}{Optional activation function to pass through after summing weights and biases of perceptrons.}

\item{...}{Additional arguments to pass to the activation function.}

\item{oprim_control}{List of control arguments to pass to \link{optim}.}
}
\value{
A list with the \link{optim} output and other arguments specified in the call.
}
\description{
A shitty wrapper for \link{optim}, which will evaluate optimization of weights and biases to fit one matrix to another using perceptrons.
}
\details{
This function optimizes weights and biases of perceptrons to map input matrix \code{x} to \code{y}. Perceptrons are densely connected to all inputs and outputs.
}
\note{
The length of \code{units} will determine the output length.
}
